requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
chardet==4.0.0
distlib==0.3.8
Django==4.2.15
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
idna==2.10
inflection==0.5.1
packaging==24.1
Pillow==10.0.0
platformdirs==4.2.2
pytz==2024.2
PyYAML==6.0.2
requests==2.25.1
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
virtualenv==20.26.3
